pyproject.toml
[tool.poetry]
name = "temelio"
version = "0.1.0"
description = ""
authors = ["Ajit Singh <ajitsinghkaler0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2"
djangorestframework = "^3.16.0"
django-background-tasks = "^1.2.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
